+++ src/Controller/TypesController.php
@@ -20,10 +20,21 @@ class TypesController extends AppController
      */
     public function index()
     {
-        $types = $this->paginate($this->Types);
+        //Check if its admin or not
+        //In our DB the id for the admin user is 11 ,so we will check that only
+        //admin user --  (username= imadmin , password = admin)
+        $ExtraInfo=$this->loadComponent('Extra_info');
+        $current_user=$ExtraInfo->getCurrentUser($this);
+        if($current_user==11){
+            $types = $this->paginate($this->Types);
 
-        $this->set(compact('types'));
-        $this->set('_serialize', ['types']);
+            $this->set(compact('types'));
+            $this->set('_serialize', ['types']);
+        }
+        //If its not admin user than redirect him
+        else{
+            $this->redirect(['controller'=>'Dashboard','action'=>'index']);
+        }
     }
 
     /**
@@ -35,12 +46,23 @@ class TypesController extends AppController
      */
     public function view($id = null)
     {
-        $type = $this->Types->get($id, [
-            'contain' => ['Menus', 'Users']
-        ]);
+        //Check if its admin or not
+        //In our DB the id for the admin user is 11 ,so we will check that only
+        //admin user --  (username= imadmin , password = admin)
+        $ExtraInfo=$this->loadComponent('Extra_info');
+        $current_user=$ExtraInfo->getCurrentUser($this);
+        if($current_user==11){
+            $type = $this->Types->get($id, [
+                'contain' => ['Menus', 'Users']
+            ]);
 
-        $this->set('type', $type);
-        $this->set('_serialize', ['type']);
+            $this->set('type', $type);
+            $this->set('_serialize', ['type']);
+        }
+        //If its not admin user than redirect him
+        else{
+            $this->redirect(['controller'=>'Dashboard','action'=>'index']);
+        }
     }
 
     /**
@@ -50,23 +72,34 @@ class TypesController extends AppController
      */
     public function add()
     {
+        //Check if its admin or not
+        //In our DB the id for the admin user is 11 ,so we will check that only
+        //admin user --  (username= imadmin , password = admin)
         $ExtraInfo=$this->loadComponent('Extra_info');
-        $type = $this->Types->newEntity();
-        if ($this->request->is('post')) {
-            $type = $this->Types->patchEntity($type, $this->request->getData());
-            $current_user=$ExtraInfo->getCurrentUser($this);
-            $ExtraInfo->setCreatedBy($type,$current_user);
-            $ExtraInfo->setModifiedBy($type,$current_user);
-            if ($this->Types->save($type)) {
-                $this->Flash->success(__('The {0} has been saved.', 'Type'));
-                return $this->redirect(['action' => 'index']);
-            } else {
-                $this->Flash->error(__('The {0} could not be saved. Please, try again.', 'Type'));
+        $current_user=$ExtraInfo->getCurrentUser($this);
+        if($current_user==11){
+            $ExtraInfo=$this->loadComponent('Extra_info');
+            $type = $this->Types->newEntity();
+            if ($this->request->is('post')) {
+                $type = $this->Types->patchEntity($type, $this->request->getData());
+                $current_user=$ExtraInfo->getCurrentUser($this);
+                $ExtraInfo->setCreatedBy($type,$current_user);
+                $ExtraInfo->setModifiedBy($type,$current_user);
+                if ($this->Types->save($type)) {
+                    $this->Flash->success(__('The {0} has been saved.', 'Type'));
+                    return $this->redirect(['action' => 'index']);
+                } else {
+                    $this->Flash->error(__('The {0} could not be saved. Please, try again.', 'Type'));
+                }
             }
+            $menus = $this->Types->Menus->find('list', ['limit' => 200]);
+            $this->set(compact('type', 'menus'));
+            $this->set('_serialize', ['type']);
         }
-        $menus = $this->Types->Menus->find('list', ['limit' => 200]);
-        $this->set(compact('type', 'menus'));
-        $this->set('_serialize', ['type']);
+        //If its not admin user than redirect him
+        else{
+            $this->redirect(['controller'=>'Dashboard','action'=>'index']);
+        }
     }
 
     /**
@@ -78,24 +111,35 @@ class TypesController extends AppController
      */
     public function edit($id = null)
     {
+        //Check if its admin or not
+        //In our DB the id for the admin user is 11 ,so we will check that only
+        //admin user --  (username= imadmin , password = admin)
         $ExtraInfo=$this->loadComponent('Extra_info');
-        $type = $this->Types->get($id, [
-            'contain' => ['Menus']
-        ]);
-        if ($this->request->is(['patch', 'post', 'put'])) {
-            $type = $this->Types->patchEntity($type, $this->request->getData());
-            $current_user=$ExtraInfo->getCurrentUser($this);
-            $ExtraInfo->setModifiedBy($type,$current_user);
-            if ($this->Types->save($type)) {
-                $this->Flash->success(__('The {0} has been saved.', 'Type'));
-                return $this->redirect(['action' => 'index']);
-            } else {
-                $this->Flash->error(__('The {0} could not be saved. Please, try again.', 'Type'));
+        $current_user=$ExtraInfo->getCurrentUser($this);
+        if($current_user==11){
+            $ExtraInfo=$this->loadComponent('Extra_info');
+            $type = $this->Types->get($id, [
+                'contain' => ['Menus']
+            ]);
+            if ($this->request->is(['patch', 'post', 'put'])) {
+                $type = $this->Types->patchEntity($type, $this->request->getData());
+                $current_user=$ExtraInfo->getCurrentUser($this);
+                $ExtraInfo->setModifiedBy($type,$current_user);
+                if ($this->Types->save($type)) {
+                    $this->Flash->success(__('The {0} has been saved.', 'Type'));
+                    return $this->redirect(['action' => 'index']);
+                } else {
+                    $this->Flash->error(__('The {0} could not be saved. Please, try again.', 'Type'));
+                }
             }
+            $menus = $this->Types->Menus->find('list', ['limit' => 200]);
+            $this->set(compact('type', 'menus'));
+            $this->set('_serialize', ['type']);
         }
-        $menus = $this->Types->Menus->find('list', ['limit' => 200]);
-        $this->set(compact('type', 'menus'));
-        $this->set('_serialize', ['type']);
+        //If its not admin user than redirect him
+        else{
+            $this->redirect(['controller'=>'Dashboard','action'=>'index']);
+        }
     }
 
     /**
@@ -107,13 +151,24 @@ class TypesController extends AppController
      */
     public function delete($id = null)
     {
-        $this->request->allowMethod(['post', 'delete']);
-        $type = $this->Types->get($id);
-        if ($this->Types->delete($type)) {
-            $this->Flash->success(__('The {0} has been deleted.', 'Type'));
-        } else {
-            $this->Flash->error(__('The {0} could not be deleted. Please, try again.', 'Type'));
+        //Check if its admin or not
+        //In our DB the id for the admin user is 11 ,so we will check that only
+        //admin user --  (username= imadmin , password = admin)
+        $ExtraInfo=$this->loadComponent('Extra_info');
+        $current_user=$ExtraInfo->getCurrentUser($this);
+        if($current_user==11){
+            $this->request->allowMethod(['post', 'delete']);
+            $type = $this->Types->get($id);
+            if ($this->Types->delete($type)) {
+                $this->Flash->success(__('The {0} has been deleted.', 'Type'));
+            } else {
+                $this->Flash->error(__('The {0} could not be deleted. Please, try again.', 'Type'));
+            }
+            return $this->redirect(['action' => 'index']);
         }
-        return $this->redirect(['action' => 'index']);
+        //If its not admin user than redirect him
+        else{
+            $this->redirect(['controller'=>'Dashboard','action'=>'index']);
+        }
     }
 }
